Moved the files to the correct team name directory
